IGNITE-26067 Review notes - remove redundant interfaces + fix javadoc params.
