+++ modules/table/src/test/java/org/apache/ignite/internal/table/distributed/PartitionModificationCounterTest.java
@@ -15,18 +15,12 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.sql.statistic;
+package org.apache.ignite.internal.table.distributed;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 
-import org.apache.ignite.internal.TestHybridClock;
-import org.apache.ignite.internal.hlc.ClockService;
 import org.apache.ignite.internal.hlc.HybridTimestamp;
-import org.apache.ignite.internal.hlc.TestClockService;
-import org.apache.ignite.internal.sql.engine.statistic.PartitionModificationCounterFactoryImpl;
-import org.apache.ignite.internal.table.distributed.PartitionModificationCounter;
-import org.apache.ignite.internal.table.distributed.PartitionModificationCounterFactory;
 import org.apache.ignite.internal.testframework.BaseIgniteAbstractTest;
 import org.apache.ignite.internal.testframework.IgniteTestUtils;
 import org.junit.jupiter.api.Test;
@@ -35,30 +29,28 @@ import org.junit.jupiter.api.Test;
  * Tests for class {@link PartitionModificationCounter}.
  */
 public class PartitionModificationCounterTest extends BaseIgniteAbstractTest {
-    private final TestHybridClock clock = new TestHybridClock(() -> 1L);
-    private final ClockService clockService = new TestClockService(clock);
     private final PartitionModificationCounterFactory factory =
-            new PartitionModificationCounterFactoryImpl(clockService);
+            new PartitionModificationCounterFactory(() -> HybridTimestamp.hybridTimestamp(1L));
 
     @Test
     void initialValues() {
         PartitionModificationCounter counter = factory.create(() -> 10_000L);
 
-        assertThat(counter.lastMilestoneTimestamp().getPhysical(), is(1L));
+        assertThat(counter.lastMilestoneTimestamp().longValue(), is(1L));
         assertThat(counter.value(), is(0L));
 
         counter.updateValue(0, HybridTimestamp.MAX_VALUE);
-        assertThat(counter.lastMilestoneTimestamp().getPhysical(), is(1L));
+        assertThat(counter.lastMilestoneTimestamp().longValue(), is(1L));
         assertThat(counter.value(), is(0L));
     }
 
     @Test
     void lastMilestoneTimestampUpdate() {
         int rowsCount = 10_000;
-        int threshold = (int) (rowsCount * PartitionModificationCounterFactoryImpl.DEFAULT_STALE_ROWS_FRACTION);
+        int threshold = (int) (rowsCount * PartitionModificationCounterFactory.DEFAULT_STALE_ROWS_FRACTION);
         PartitionModificationCounter counter = factory.create(() -> rowsCount);
 
-        assertThat(counter.lastMilestoneTimestamp().getPhysical(), is(1L));
+        assertThat(counter.lastMilestoneTimestamp().longValue(), is(1L));
 
         {
             HybridTimestamp commitTime = HybridTimestamp.hybridTimestamp(100L);
